More AssetListController tests
Started Testing search-method

+++ AMS/Controllers/AssetListController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using AMS.Controllers.Interfaces;
 using AMS.Database.Repositories.Interfaces;
@@ -13,7 +14,7 @@ namespace AMS.Controllers
     {
         private IExporter _exporter;
         private IAssetRepository _assetRepository;
-        
+
         public List<Asset> AssetList { get; set; }
         public List<Tag> TagsList { get; set; } 
 
@@ -24,41 +25,71 @@ namespace AMS.Controllers
             _exporter = exporter;
 
         }
+        
+        /// <summary>
+        /// Filters the list if Assets to only contain assets that match the searchQuery
+        /// </summary>
+        /// <param name="query"></param>
         public void Search(string query)
         {
             //TODO: Filter list based on search query, also search by tags
-            throw new NotImplementedException();
+            List<Asset> searchResult = _assetRepository.Search(query).ToList();
+            AssetList = searchResult;
         }
 
+        /// <summary>
+        /// Displays the editPage for a blank asset
+        /// </summary>
         public void AddNew()
         {
             //TODO: Redirect to blank editPage
             AssetList.Add(new Asset());
         }
 
+        /// <summary>
+        /// Displays the editPage for the given page
+        /// </summary>
+        /// <param name="asset"></param>
+        /// <exception cref="NotImplementedException"></exception>
         public void Edit(Asset asset)
         {
             //TODO: Redirect to EditPage
             throw new NotImplementedException();
         }
         
+        /// <summary>
+        /// Displays the viewPage for the given asset
+        /// </summary>
+        /// <param name="asset"></param>
+        /// <exception cref="NotImplementedException"></exception>
         public void ViewAsset(Asset asset)
         {
             //TODO: Redirect to ViewPage
             throw new NotImplementedException();
         }
 
+        /// <summary>
+        /// Removes the given asset from the database and from the list
+        /// </summary>
+        /// <param name="asset"></param>
         public void Remove(Asset asset)
         {
+            // Check if asset is in list
+            if (!AssetList.Contains(asset))
+                return;
             // Delete the asset
             if (_assetRepository.Delete(asset))
                 AssetList.Remove(asset);
         }
 
+        /// <summary>
+        /// Exports the selected items in the listView to a .csv file
+        /// </summary>
+        /// <param name="assets"></param>
         public void Export(List<Asset> assets)
         {
             //TODO: Implement printHelper
-            throw new NotImplementedException();
+            _exporter.ExportSelected();
         }
     }
 }
